shared useP5DefaultSetup fn
